refactor: Extract the generic download engine into DownloadService
## Summary
- Split the streaming download machinery out of IPSWService into a new
  DownloadService so future non-IPSW downloads (e.g. Linux installer ISOs)
  can reuse it. No behavior change.
- IPSWService shrinks to the macOS-specific surface: latest-restore-image
  lookup plus thin delegation into the DownloadService it owns. Its public
  init and IPSWProviding conformance are unchanged.

## Changes
- New Kernova/Services/DownloadService.swift: resumable .kernovadownload
  transfer engine (200/206/416 handling, Range/If-Range resume, EWMA
  progress, per-destination static claim table), DownloadBundle,
  DownloadBundleMetadata, DownloadError, StreamingDataTaskDelegate,
  DownloadSpeedSmoother.
- DownloadProgress moved from MacOSInstallState.swift to its own model file.
- IPSWService.swift reduced to fetchLatestRestoreImage + delegation;
  IPSWError keeps only .noDownloadURL; IPSWService becomes a Sendable
  struct now that the URLSession lifetime lives in DownloadService.
- Call sites retargeted: VMCreationViewModel, VMInstance+Display,
  VMLifecycleCoordinator, RestoreImageProbeService doc cross-reference.
- Tests: IPSWServiceDownloadTests/IPSWBundleTests move near-verbatim to
  DownloadServiceTests/DownloadBundleTests; new IPSWServiceTests covers the
  delegation path without touching the shared StubURLProtocol handler.
- docs/ARCHITECTURE.md: DownloadService entry added, IPSWService entry
  rewritten, FileSystemOperating seam holder corrected.

## Test plan
- [x] make test — 2172 passed, 0 failed (full suite)
- [x] make lint — clean
